requirements.txt
appdirs==1.4.3
attrs==19.3.0
autopep8==1.4.4
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
comet-git-pure==0.19.16
comet-ml==3.1.7
configobj==5.0.6
distlib==0.3.0
entrypoints==0.3
everett==1.0.2
filelock==3.0.12
flake8==3.7.9
future==0.18.2
GPUtil==1.4.0
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
joblib==0.14.1
jsonschema==3.0.2
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mypy-extensions==0.4.3
netifaces==0.10.9
numpy==1.18.1
nvidia-ml-py3==7.352.0
olefile==0.46
pathspec==0.7.0
Pillow==7.1.2
pipenv==2018.11.26
psutil==5.7.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyrsistent==0.16.0
regex==2020.4.4
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
thop==0.0.31.post2005141830
toml==0.10.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
websocket-client==0.57.0
wincertstore==0.2
wurlitzer==2.0.0
zipp==3.1.0
